+++ src/driver/mysql/MysqlQueryRunner.ts
@@ -1,9 +1,7 @@
 import {QueryRunner} from "../../query-runner/QueryRunner";
-import {DatabaseConnection} from "../DatabaseConnection";
 import {ObjectLiteral} from "../../common/ObjectLiteral";
 import {TransactionAlreadyStartedError} from "../error/TransactionAlreadyStartedError";
 import {TransactionNotStartedError} from "../error/TransactionNotStartedError";
-import {DataTypeNotSupportedByDriverError} from "../error/DataTypeNotSupportedByDriverError";
 import {ColumnSchema} from "../../schema-builder/schema/ColumnSchema";
 import {ColumnMetadata} from "../../metadata/ColumnMetadata";
 import {TableSchema} from "../../schema-builder/schema/TableSchema";
@@ -11,10 +9,9 @@ import {ForeignKeySchema} from "../../schema-builder/schema/ForeignKeySchema";
 import {PrimaryKeySchema} from "../../schema-builder/schema/PrimaryKeySchema";
 import {IndexSchema} from "../../schema-builder/schema/IndexSchema";
 import {QueryRunnerAlreadyReleasedError} from "../../query-runner/error/QueryRunnerAlreadyReleasedError";
-import {ColumnType} from "../types/ColumnTypes";
 import {Connection} from "../../connection/Connection";
 import {MysqlConnectionOptions} from "./MysqlConnectionOptions";
-import {ColumnOptions} from "../../decorator/options/ColumnOptions";
+import {MysqlDriver} from "./MysqlDriver";
 
 /**
  * Runs queries on a single mysql database connection.
@@ -36,12 +33,16 @@ export class MysqlQueryRunner implements QueryRunner {
      */
     isTransactionActive = false;
 
+    /**
+     * Real database connection from a connection pool used to perform queries.
+     */
+    databaseConnection: any;
+
     // -------------------------------------------------------------------------
     // Constructor
     // -------------------------------------------------------------------------
 
-    constructor(protected connection: Connection,
-                protected databaseConnection: DatabaseConnection) {
+    constructor(protected connection: Connection) {
     }
 
     // -------------------------------------------------------------------------
@@ -49,16 +50,26 @@ export class MysqlQueryRunner implements QueryRunner {
     // -------------------------------------------------------------------------
 
     /**
+     * Creates/uses connection from the connection pool to perform further operations.
+     */
+    connect(): Promise<any> {
+        return new Promise((ok, fail) => {
+            const driver = this.connection.driver as MysqlDriver;
+            driver.pool.getConnection((err: any, connection: any) => {
+                this.databaseConnection = connection;
+                err ? fail(err) : ok(connection);
+            });
+        });
+    }
+
+    /**
      * Releases database connection. This is needed when using connection pooling.
      * If connection is not from a pool, it should not be released.
      * You cannot use this class's methods after its released.
      */
     release(): Promise<void> {
-        if (this.databaseConnection.releaseCallback) {
-            this.isReleased = true;
-            return this.databaseConnection.releaseCallback();
-        }
-
+        this.isReleased = true;
+        this.databaseConnection.release();
         return Promise.resolve();
     }
 
@@ -144,7 +155,7 @@ export class MysqlQueryRunner implements QueryRunner {
 
         return new Promise((ok, fail) => {
             this.connection.logger.logQuery(query, parameters);
-            this.databaseConnection.connection.query(query, parameters, (err: any, result: any) => {
+            this.databaseConnection.query(query, parameters, (err: any, result: any) => {
                 if (err) {
                     this.connection.logger.logFailedQuery(query, parameters);
                     this.connection.logger.logQueryError(err);